research-agent/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-agent"
version = "0.1.0"
description = "An AI agent that researches topics and builds long-term memory across sessions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Research Agent Team"}
]

dependencies = [
    "google-genai>=0.1.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "tavily-python>=0.3.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
